Address human-reg PR review structure
Move LBF-specific BC helpers out of the generic BC agent package and keep executable experiment entrypoints under scripts.

Constraint: PR 89 review requested scripts under scripts/ and LBF BC initialization infrastructure under agents/lbf/bc.

Rejected: Keep BC rollout scripts in agents/bc | the agents package should only expose agent implementations and architectures.

Confidence: medium

Scope-risk: moderate

Directive: Keep future LBF BC-specific helpers under agents/lbf/bc or scripts/bc instead of agents/bc.

Tested: python3 -m compileall agents/bc agents/lbf scripts/bc ego_agent_training/ppo_ego.py marl/ippo.py tests/test_lbf_human_reg_features.py; git diff --check

Not-tested: pytest and direct JAX runtime checks on this Mac; local Python has no pytest or jax installed.
