+++ agents/bc/__init__.py
@@ -2,5 +2,4 @@ from agents.bc.bc_lstm import (
     BCLSTMAgent,
     BCLSTMConfig,
     BCLSTMNetwork,
-    BCLSTMPolicyWrapper,
 )
