+++ agents/lbf/bc/__init__.py
@@ -1,0 +1,13 @@
+from agents.lbf.bc.configs import (
+    LBF_CONFIGS,
+    load_bc_config,
+    resolve_first_existing_path,
+)
+from agents.lbf.bc.features import augment_lbf_obs
+
+__all__ = [
+    "LBF_CONFIGS",
+    "augment_lbf_obs",
+    "load_bc_config",
+    "resolve_first_existing_path",
+]
