+++ scripts/bc/__init__.py
@@ -1,0 +1,1 @@
+"""BC training and evaluation scripts."""
