+++ scripts/bc/run_human_reg_ppo_sweep.py
@@ -12,7 +12,7 @@ REPO_ROOT = Path(__file__).resolve().parents[2]
 if str(REPO_ROOT) not in sys.path:
     sys.path.insert(0, str(REPO_ROOT))
 
-from agents.bc.evaluate_lbf import LBF_CONFIGS
+from agents.lbf.bc import LBF_CONFIGS
 from agents.lbf.agent_policy_wrappers import (
     LBFEntitledPolicyWrapper,
     LBFGreedyHeuristicPolicyWrapper,
